Add documentation for the dynamic routing interfaces.  Change Ipv4::AddRoutingProtocol priority parameter type from int to int16_t to make way for in the future providing stable sorting of routing protocols of the same priority.

+++ src/internet-node/ipv4-impl.h
@@ -36,7 +36,7 @@ public:
   virtual ~Ipv4Impl ();
 
   virtual void AddRoutingProtocol (Ptr<Ipv4RoutingProtocol> routingProtocol,
-                                   int priority);
+                                   int16_t priority);
 
   virtual void AddHostRouteTo (Ipv4Address dest, 
 			       Ipv4Address nextHop,
